refactor(executors): extract pure upstream-header helpers from base (1539->1451)
Port upstream #6008. Pure code motion into open-sse/executors/base/headers.ts leaf; base.ts re-imports. Verified no fork-only logic reverted (delete-hunk guard). Regression guard: tests/unit/base-headers-split.test.ts.
